requirements.txt
langchain>=1.0.0
langchain-openai>=1.0.0
pandas>=2.2.0
python-dotenv>=1.0.0
